requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
base58==2.1.1
binaryornot==0.4.4
bitarray==2.6.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
cytoolz==0.12.1
Django==4.1.5
django-browser-reload==1.6.0
django-tailwind==3.4.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.22.4
pandas==1.5.3
parsimonious==0.8.1
protobuf==3.19.5
pycryptodome==3.16.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
rlp==2.0.1
RPi.GPIO==0.7.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toolz==0.12.0
urllib3==1.26.13
varint==1.0.2
web3==5.31.3
websockets==9.1
yarl==1.8.2
